fix storage

+++ helpers/storage.js
@@ -1,5 +1,5 @@
 import AsyncStorage from "@react-native-async-storage/async-storage";
-
+import { Platform } from "react-native";
 export const setItem = (key, value) =>
   AsyncStorage.setItem(key, JSON.stringify(value));
 
@@ -8,6 +8,14 @@ export const getItem = async (key) => {
   return JSON.parse(object);
 };
 
-export const clearStorage = () => {
-  AsyncStorage.clear();
+export const clearStorage = async () => {
+  const asyncStorageKeys = await AsyncStorage.getAllKeys();
+  if (asyncStorageKeys.length > 0) {
+    if (Platform.OS === "android") {
+      await AsyncStorage.clear();
+    }
+    if (Platform.OS === "ios") {
+      await AsyncStorage.multiRemove(asyncStorageKeys);
+    }
+  }
 };
